add asymmetricSimdParser for shapes (B,D,K) + (D,K)
