+++ stream/workload/flatten_node.py
@@ -2,7 +2,7 @@ import numpy as np
 from zigzag.datatypes import LayerOperand
 from zigzag.workload.layer_node_abc import LayerNodeABC
 
-from stream.utils import NodeTensor
+from stream.node_tensor import NodeTensor
 from stream.workload.node import Node
